LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-insert-br-expected.txt


FAIL pointer updates (adding br elements) assert_equals: error events on c expected 1 but got 0
